// src/types/type.ts
export interface CustomEventInterface {

  type: string

  phase: number

  ns?: string

  target?: any

  originalEvent?: CustomEventInterface | Event

  isPrevented?: true

  isStopped?: true

  listener?: Function

  /**
   * 阻止事件的默认行为
   */
  preventDefault(): this;

  /**
   * 停止事件广播
   */
  stopPropagation(): this;

  prevent(): this;

  stop(): this;

}